Fix determination of remote PIDs
Since the automatix allows multiline and more complex commands checking
the process list for the whole command seems to be a bad idea. Depending
on the command the PIDs returned might contain totally wrong and
possibly dangerous process IDs.
We now check for the RUNNING_INSIDE_AUTOMATIX variable which is
contained in all remote commands. Also there should be only 1 pid with
this string. Otherwise we now give a warning.